Health Files now moved to correct folders - building lib-tempo
