+++ src/lib-tempo/include/tempo/entity/EntityCreationClient.hpp
@@ -6,13 +6,13 @@
 #include <Ogre.h>
 
 #include <tempo/component/ComponentGridAi.hpp>
+#include <tempo/component/ComponentHealth.hpp>
 #include <tempo/entity/EntityCreation.hpp>
 #include <tempo/entity/Render.hpp>
 #include <tempo/entity/ID.hpp>
 #include <tempo/entity/LevelManager.hpp>
 #include <tempo/entity/PlayerLocal.hpp>
 #include <tempo/entity/PlayerRemote.hpp>
-#include <tempo/entity/ComponentHealth.hpp>
 
 namespace tempo
 {
